utils: initial local DB for openQA
Per discussion with Marek a local DB is best suited for doing
queries on the testfailures than the previous (and still present)
implementation of multiple "filter functions".

Considerations against using just openQA's DB:
  - the external test resutls ("# system-out:...) which are needed
for analysis are stored as files instead.

The end goal is to have more flexibility in querying the test data
to find instability patterns.