utils: avoid recusive db object creation
There was a recursion bug because we were trying to create the
children for a parent job and the parents for the child job and
only committing that in the end.

This would lead to the double insertion of jobs. Now we flush
objects when they are created, which means they will be found via
the queries and thus, not be re-created.