utils: determine instability based on the updates branch
Per Marek's request the instability is to be determined from the
updates branch. This will ensure the stability data is not
influenced by faulty pull requests.

If the report is being provided to a non-updates job, then to make
this work we try to find a job in updates whose ID is the closes
to the job passed via --job-id. This should ensure temporal
proximity to that job and thus more accurate stability data.

Otherwise we running github_reporting would not be deterministic
since it would depend on the time in which it was ran.