Increase few timeouts
And also add workaround for selecting wrong application in VM settings: when an
entry near top is clicked, the selection list may scroll a bit (1 position in
this case) which results in later clicking wrong entry. Fix this by re-checking
if the correct entry is selected and clicking again if it isn't (this time -
using the position after the scroll).