Allow choosing test flavor for PRs test
This is useful to run PRs test (multiple PRs at once) with the kernel test
flavor - for example to test S0ix.
This means in some cases the kernel test flavor will have updates for VMs too -
do not disable those unconditionally anymore, but test if there is something in
the repository. For now hardcode checking for bookworm specifically (should be
good for quite some time), but it will need a more generic solution at some point.