Improve handling LUKS prompt on hardware runners
On laptop runners, LUKS prompt is not visible on HDMI output by default. It
shows up only after pressing ESC twice.
Improve handling by looking for plymouth text output, instead of a timeout on a
LUKS prompt - this speeds up startup in most cases. But unfortunately runners
with Heads (or with old Intel GPU?) do not show anything on HDMI in that case
and waiting for a blank screen isn't really what is needed here. In that case,
keep the timeout approach.