Remaining parts to make sys-gui-gpu tests working
- check for GUIVM_GPU in few more places
- set root password - used on text console
- start sys-gui-gpu in proper place
- add xe.force_probe=a780 kernel opt, as i915 driver doesn't work with passthrough on the hw13 worker

The second point isn't strictly correct, as the test expect to login to dom0,
not sys-gui-gpu. It wants to execute fstrim and then shutdown (if disk image is
going to be saved). But fortunately, it doesn't matter for this test, as disk
image is not going to be saved, so just successful login is enough to make test
pass.