Fix qubesteststub for PCI path
Don't use `lspci` output as devices don't match qvm-pci ident anymore. Discover
devices via qubes API. That can't be done in the extension __init__ as it
doesn't have the Qubes() reference yet, do that in domain-init event. This
makes the structure a bit weird, as the old stub (that uses lspci still) does
it still in __init__... This will be "fixed" once 4.2 goes EOL in a year or so.