Install prod keyring package in dev and staging
The final RPM (regardless of the environment) requires the upstream
keyring package (see [1]). Since this is using the securedrop dev
environment, it can't obtain it. To solve this, we're installing the
production keyring package through Qubes-Contrib, just as a regular user
would.

The key bit to make this work is that the dev and staging packages will
always have higher version numbers than the production repositories. Due
to this fact, the packages installed will be the ones for the
environment we're in.

[1]: https://github.com/freedomofpress/securedrop-workstation/pull/1450/files#diff-5d2e11f87f29c1289d107de90bd520aa145edde09453d280d1c446e84b7c3aa0R43