API: scheduling multi-PR runs via github comments and filtering
Allow specifying `TEST` variable when scheduling tests, which openQA
will use to schedule subset of tests. Similarly, allow specifying
`MACHINE` but here do a little custom processing - openQA supports just
a single value, not a list, so handle lists internally and schedule each
selected value separately (but with the same `BUILD` value, so they are
grouped together).

On top of that, when `openQArun` comment is called with `PR_LABEL`
parameter, schedule build of PRs with that label on gitlab. This used to
be done on gitlab directly, but that requires extra permissions on
gitlab which are not flexible enough.
Note in this case, it doesn't matter anymore on which PR the comment is
posted, it could be even one without that label. This might be changed
in the future.